Switches between db type (pgsql, mssql)
OpenIMIS works with two types of database: PostgreSQL and Microsoft SQL Server.
To ease the switch between both, a new command `db` allows to manage the
settings and the command to run the right type of database.

This change required a few updates and additions such as the generation of
the dotenv file, or its update to reflect the right config.