Prepare the Msft SQL Server correctly
After investigation it appears that the Mssql DB was not prepared correctly.
Some procedure wasn't updated for instance (e.g. `uspAPIGetClaim`). This is
because they retrieve the scripts from the last release. If you are on the
develop branch, this is not the right one.

To fix it, we mount a local directory where we put the right scripts and we
clean and prepared the DB with it.