Fix the build of the backend image and update with last dist dkr version
I discovered that the build was broken and worked only because I did it
manually. The image being there, there weren't anything to build. It didn't work
for serveral reasons (Dockerfile in a parent directory of the context, and
syntax).

Once fixed, I improved it so we would have two flavors for the same image, one
for each database. Also it was made sure that database volume is dropped when
switching to another one in order to avoid conflict.

This also includes a few changes to work with the last version of dist-dkr.