Update the environment to work with the new release
The recent release of end of May/beginning of June has introduced
several changes such as using the frontend service rather than the
gateway to route to other services, the services definition, or the
usage of the recent releases of the backend.

We discovered a few issues related to that and updated the dev
environment such as the ineffective removal of the database due to
existing connection (probably the healthcheck) or a failing autostart
of the scheduler backend.

The present changes makes sure we can work with the new release and
fix some other unrelated bugs we didn't discover before (unproper way
to use service image name in Docker compose manifest, backend command,
not idempotent function to download mssql scripts).